+++ source/WesternStatesWater.WaDE.Accessors/EntityFramework/AllocationAmountsFact.cs
@@ -33,7 +33,7 @@ namespace WesternStatesWater.WaDE.Accessors.EntityFramework
         public double? AllocationCropDutyAmount { get; set; }
         public double? AllocationAmount { get; set; }
         public double? AllocationMaximum { get; set; }
-        public double? PopulationServed { get; set; }
+        public long? PopulationServed { get; set; }
         public double? PowerGeneratedGwh { get; set; }
         public double? IrrigatedAcreage { get; set; }
         public string AllocationCommunityWaterSupplySystem { get; set; }
